Move runtime data model into init (#623)
